RSDK-4388-nick-proposal
