+++ services/motion/builtin/builtin.go
@@ -7,14 +7,12 @@ import (
 	"fmt"
 	"math"
 	"sync"
-	"time"
 
 	"github.com/edaniels/golog"
 	"github.com/golang/geo/r3"
 	geo "github.com/kellydunn/golang-geo"
 	"github.com/pkg/errors"
 	servicepb "go.viam.com/api/service/motion/v1"
-	goutils "go.viam.com/utils"
 
 	"go.viam.com/rdk/components/base"
 	"go.viam.com/rdk/components/base/kinematicbase"
@@ -239,16 +237,13 @@ func (ms *builtIn) MoveOnGlobe(
 	ctx context.Context,
 	componentName resource.Name,
 	destination *geo.Point,
-	heading float64,
+	_heading float64,
 	movementSensorName resource.Name,
 	obstacles []*spatialmath.GeoObstacle,
 	motionCfg *motion.MotionConfiguration,
 	extra map[string]interface{},
 ) (bool, error) {
 	operation.CancelOtherWithLabel(ctx, builtinOpLabel)
-
-	positionPollingPeriod := time.Duration(1000/motionCfg.PositionPollingFreqHz) * time.Millisecond
-	obstaclePollingPeriod := time.Duration(1000/motionCfg.ObstaclePollingFreqHz) * time.Millisecond
 
 	movementSensor, ok := ms.movementSensors[movementSensorName]
 	if !ok {
@@ -260,283 +255,40 @@ func (ms *builtIn) MoveOnGlobe(
 		return false, err
 	}
 
-	// successChan is for sending messages that relate to the base meeting the success criteria for the request
-	successChan := make(chan bool, 1)
-	defer close(successChan)
-
-	// replanChan is for sending messages that signal that we should stop motion and replan
-	// initialize with a true signal so that we immediately start planning
-	replanChan := make(chan bool, 1)
-	defer close(replanChan)
-	replanChan <- true
-
-	// errChan is for sending error messages encountered during the move
-	errChan := make(chan error)
-	defer close(errChan)
-
-	// create a waitgroup to manage goroutines and wait on their completion before exiting the function
-	var backgroundWorkers sync.WaitGroup
-	defer backgroundWorkers.Wait()
-
-	// create a cancellable context and ensure that it is cancelled before exiting
-	var cancelCtx context.Context
-	var cancelFn context.CancelFunc
-	defer func() {
-		if cancelFn != nil {
-			cancelFn()
-		}
-	}()
-
-	// helper function to manage polling functions
-	startPolling := func(ctx context.Context, period time.Duration, fn func(context.Context) error) {
-		backgroundWorkers.Add(1)
-		goutils.ManagedGo(func() {
-			ticker := time.NewTicker(period)
-			defer ticker.Stop()
-			for {
-				select {
-				case <-ctx.Done():
-					return
-				case <-ticker.C:
-					if err := fn(ctx); err != nil {
-						errChan <- err
-					}
-					return
-				}
-			}
-		}, backgroundWorkers.Done)
+	plan, err := plan(ctx, planRequest, kb, componentName.Name)
+	if err != nil {
+		return false, err
 	}
+
+	planManager := newExecutionManager(ctx, plan)
+	spawnExecute(motionCfg, kb, ms, movementSensor, destination, planManager)
 
 	// start the loop that (re)plans when something is read from the replan channel
 	// and exits when something is read from the success channel
 	for {
 		select {
 		case <-ctx.Done():
+			planManager.flush()
 			return false, ctx.Err()
-		case <-successChan:
+		case <-planManager.successChan:
+			planManager.flush()
 			return true, nil
-		case err := <-errChan:
+		case err := <-planManager.executionErrChan:
+			planManager.flush()
 			return false, err
-		case <-replanChan:
+		case err := <-planManager.obsticlePollingErrChan:
+			planManager.flush()
+			return false, err
+		case err := <-planManager.positionPollingErrChan:
+			planManager.flush()
+			return false, err
+		case <-planManager.replanChan:
 			// cancel the goroutines spawned by this function, create a new cancellable context to use in the future
-			if cancelFn != nil {
-				cancelFn()
-			}
-			backgroundWorkers.Wait()
-			// context is not lost because it is cancelled above and then cancelled again with defer
-			//nolint:govet
-			cancelCtx, cancelFn = context.WithCancel(ctx)
-
-			inputs, err := kb.CurrentInputs(ctx)
-			if err != nil {
-				//nolint:govet
-				return false, err
-			}
-			// TODO: this is really hacky and we should figure out a better place to store this information
-			if len(kb.Kinematics().DoF()) == 2 {
-				inputs = inputs[:2]
-			}
-			planRequest.StartConfiguration = map[string][]referenceframe.Input{componentName.Name: inputs}
-
-			plan, err := motionplan.PlanMotion(ctx, planRequest)
-			if err != nil {
-				return false, err
-			}
-
-			// drain channels of any extra messages
-			for len(replanChan) > 0 {
-				<-replanChan
-			}
-			for len(errChan) > 0 {
-				<-errChan
-			}
-
-			// spawn two goroutines that each have the ability to trigger a replan
-			// TODO: optionally could make MotionConfiguration more robust such that the option that these are zero is never possible
-			if positionPollingPeriod > 0 {
-				startPolling(cancelCtx, positionPollingPeriod, func(ctx context.Context) error {
-					// TODO: the function that actually monitors position
-					return nil
-				})
-			}
-			if obstaclePollingPeriod > 0 {
-				startPolling(cancelCtx, obstaclePollingPeriod, func(ctx context.Context) error {
-					// TODO: the function that actually monitors obstacles
-					replanChan <- true
-					return nil
-				})
-			}
-
-			// spawn function to execute the plan on the robot
-			backgroundWorkers.Add(1)
-			goutils.ManagedGo(func() {
-				if err := ms.executePlan(cancelCtx, kb, plan); err != nil {
-					errChan <- err
-					return
-				}
-
-				// the plan has been fully executed so check to see if the GeoPoint we are at is close enough to the goal.
-				success, err := arrivedAtGoal(cancelCtx, movementSensor, destination, motionCfg.PlanDeviationMM)
-				if err != nil {
-					errChan <- err
-					return
-				}
-				if success {
-					successChan <- true
-					return
-				}
-			}, backgroundWorkers.Done)
+			planManager.flush()
+			planManager = newExecutionManager(ctx, plan)
+			spawnExecute(motionCfg, kb, ms, movementSensor, destination, planManager)
 		}
 	}
-}
-
-func arrivedAtGoal(ctx context.Context, ms movementsensor.MovementSensor, destination *geo.Point, radiusMM float64) (bool, error) {
-	position, _, err := ms.Position(ctx, nil)
-	if err != nil {
-		return false, err
-	}
-	if spatialmath.GeoPointToPose(position, destination).Point().Norm() <= radiusMM {
-		return true, nil
-	}
-	return false, nil
-}
-
-func (ms *builtIn) newMoveOnGlobeRequest(
-	ctx context.Context,
-	componentName resource.Name,
-	destination *geo.Point,
-	movementSensor movementsensor.MovementSensor,
-	obstacles []*spatialmath.GeoObstacle,
-	motionCfg *motion.MotionConfiguration,
-	extra map[string]interface{},
-) (*motionplan.PlanRequest, kinematicbase.KinematicBase, error) {
-	kinematicsOptions := kinematicbase.NewKinematicBaseOptions()
-	if motionCfg.LinearMPerSec != 0 {
-		kinematicsOptions.LinearVelocityMMPerSec = motionCfg.LinearMPerSec * 1000
-	}
-	if motionCfg.AngularDegsPerSec != 0 {
-		kinematicsOptions.AngularVelocityDegsPerSec = motionCfg.AngularDegsPerSec
-	}
-	if motionCfg.PlanDeviationMM != 0 {
-		kinematicsOptions.PlanDeviationThresholdMM = motionCfg.PlanDeviationMM
-	}
-	kinematicsOptions.GoalRadiusMM = motionCfg.PlanDeviationMM
-	kinematicsOptions.HeadingThresholdDegrees = 8
-
-	// build the localizer from the movement sensor
-	origin, _, err := movementSensor.Position(ctx, nil)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// add an offset between the movement sensor and the base if it is applicable
-	baseOrigin := referenceframe.NewPoseInFrame(componentName.ShortName(), spatialmath.NewZeroPose())
-	movementSensorToBase, err := ms.fsService.TransformPose(ctx, baseOrigin, movementSensor.Name().ShortName(), nil)
-	if err != nil {
-		movementSensorToBase = baseOrigin
-	}
-	localizer := motion.NewMovementSensorLocalizer(movementSensor, origin, movementSensorToBase.Pose())
-
-	// convert destination into spatialmath.Pose with respect to where the localizer was initialized
-	goal := spatialmath.GeoPointToPose(destination, origin)
-
-	// convert GeoObstacles into GeometriesInFrame with respect to the base's starting point
-	geoms := spatialmath.GeoObstaclesToGeometries(obstacles, origin)
-
-	gif := referenceframe.NewGeometriesInFrame(referenceframe.World, geoms)
-	worldState, err := referenceframe.NewWorldState([]*referenceframe.GeometriesInFrame{gif}, nil)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// construct limits
-	straightlineDistance := goal.Point().Norm()
-	if straightlineDistance > maxTravelDistance {
-		return nil, nil, fmt.Errorf("cannot move more than %d kilometers", int(maxTravelDistance*1e-6))
-	}
-	limits := []referenceframe.Limit{
-		{Min: -straightlineDistance * 3, Max: straightlineDistance * 3},
-		{Min: -straightlineDistance * 3, Max: straightlineDistance * 3},
-		{Min: -2 * math.Pi, Max: 2 * math.Pi},
-	}
-
-	if extra != nil {
-		if profile, ok := extra["motion_profile"]; ok {
-			motionProfile, ok := profile.(string)
-			if !ok {
-				return nil, nil, errors.New("could not interpret motion_profile field as string")
-			}
-			kinematicsOptions.PositionOnlyMode = motionProfile == motionplan.PositionOnlyMotionProfile
-		}
-	}
-	ms.logger.Debugf("base limits: %v", limits)
-
-	// create a KinematicBase from the componentName
-	baseComponent, ok := ms.components[componentName]
-	if !ok {
-		return nil, nil, resource.NewNotFoundError(componentName)
-	}
-	b, ok := baseComponent.(base.Base)
-	if !ok {
-		return nil, nil, fmt.Errorf("cannot move component of type %T because it is not a Base", baseComponent)
-	}
-
-	kb, err := kinematicbase.WrapWithKinematics(ctx, b, ms.logger, localizer, limits, kinematicsOptions)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// create a new empty framesystem which we add the kinematic base to
-	fs := referenceframe.NewEmptyFrameSystem("")
-	kbf := kb.Kinematics()
-	if err := fs.AddFrame(kbf, fs.World()); err != nil {
-		return nil, nil, err
-	}
-
-	// TODO(RSDK-3407): this does not adequately account for geometries right now since it is a transformation after the fact.
-	// This is probably acceptable for the time being, but long term the construction of the frame system for the kinematic base should
-	// be moved under the purview of the kinematic base wrapper instead of being done here.
-	offsetFrame, err := referenceframe.NewStaticFrame("offset", movementSensorToBase.Pose())
-	if err != nil {
-		return nil, nil, err
-	}
-	if err := fs.AddFrame(offsetFrame, kbf); err != nil {
-		return nil, nil, err
-	}
-
-	return &motionplan.PlanRequest{
-		Logger:             ms.logger,
-		Goal:               referenceframe.NewPoseInFrame(referenceframe.World, goal),
-		Frame:              offsetFrame,
-		FrameSystem:        fs,
-		StartConfiguration: referenceframe.StartPositions(fs),
-		WorldState:         worldState,
-		Options:            extra,
-	}, kb, nil
-}
-
-func (ms *builtIn) executePlan(ctx context.Context, kinematicBase kinematicbase.KinematicBase, plan motionplan.Plan) error {
-	waypoints, err := plan.GetFrameSteps(kinematicBase.Name().Name)
-	if err != nil {
-		return err
-	}
-
-	for i := 1; i < len(waypoints); i++ {
-		select {
-		case <-ctx.Done():
-			return nil
-		default:
-			ms.logger.Info(waypoints[i])
-			if err := kinematicBase.GoToInputs(ctx, waypoints[i]); err != nil {
-				// If there is an error on GoToInputs, stop the component if possible before returning the error
-				if stopErr := kinematicBase.Stop(ctx, nil); stopErr != nil {
-					return errors.Wrap(err, stopErr.Error())
-				}
-				return err
-			}
-		}
-	}
-	return nil
 }
 
 func (ms *builtIn) GetPose(
@@ -553,7 +305,7 @@ func (ms *builtIn) GetPose(
 		ctx,
 		referenceframe.NewPoseInFrame(
 			componentName.ShortName(),
-			spatialmath.NewPoseFromPoint(r3.Vector{0, 0, 0}),
+			spatialmath.NewPoseFromPoint(r3.Vector{X: 0, Y: 0, Z: 0}),
 		),
 		destinationFrame,
 		supplementalTransforms,
